+++ Cobalt/Cobalt.csproj
@@ -286,20 +286,6 @@
     <EmbeddedResource Include="Resources\Export\config\maps\mvm_rottenburg.xml" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Cobalt.ItemSchema.Dll\Cobalt.TF2Items.Dll.csproj">
-      <Project>{E762D8F4-DAC2-45FB-8877-0C1D6698A176}</Project>
-      <Name>Cobalt.TF2Items.Dll</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Cobalt.PopulationFile.Dll\Cobalt.Core.PopulationFile.Dll.csproj">
-      <Project>{293a2d00-73be-4e78-896d-a99620bf02ee}</Project>
-      <Name>Cobalt.Core.PopulationFile.Dll</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Cobalt.Translation.Dll\Cobalt.Core.Translation.Dll.csproj">
-      <Project>{b77daf19-498a-4c14-af6d-978736c30170}</Project>
-      <Name>Cobalt.Core.Translation.Dll</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
     <EmbeddedResource Include="Resources\Export\translation\default.xml" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
