+++ TF2/TF2.Info/Properties/AssemblyInfo.cs
