+++ TF2/TF2.Info/TF2.Info.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{66A8409E-243D-462B-A6D0-DE782B6FABA7}</ProjectGuid>
+    <ProjectGuid>{F5669172-7996-4525-82DF-D1EA104C79EB}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Valve.Keyvalue</RootNamespace>
-    <AssemblyName>valve.fileio</AssemblyName>
+    <RootNamespace>TF2.Items</RootNamespace>
+    <AssemblyName>tf2</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -40,10 +40,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="KeyValue.cs" />
-    <Compile Include="KVNode.cs" />
-    <Compile Include="KVFile.cs" />
+    <Compile Include="ItemsInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TFAttribute.cs" />
+    <Compile Include="TFItem.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\TF2.ItemSchema\TF2.ItemSchema.csproj">
+      <Project>{BB0EEF20-A25E-499A-B360-0C80E0F651A4}</Project>
+      <Name>Cobalt.TF2Info.Dll</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
