+++ TF2/TF2.Info/TFClass.cs
