+++ TF2/TF2.Info/TFSlot.cs
