+++ TF2/TF2.ItemSchema/ItemsInfo.cs
