+++ TF2/TF2.ItemSchema/Properties/AssemblyInfo.cs
