+++ TF2/TF2.ItemSchema/TFAttribute.cs
