+++ TF2/TF2.ItemSchema/TFItem.cs
