+++ TF2/TF2.Population/Elements/TFBot.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Flowaria.Population.Element
+namespace TF2.Population.Element
 {
     public class TFBot
     {
