+++ TF2/TF2.Population/Elements/WaveSpawn.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Flowaria.Population.Element
+namespace TF2.Population.Element
 {
-    public class Mission
+    public class WaveSpawn
     {
     }
 }
