+++ Valve/Valve.KeyValue/KVFile.cs
@@ -4,19 +4,20 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Valve.Keyvalue
+namespace Valve.KeyValue
 {
     public static class KVFile
     {
         public static KeyValue ImportKeyValue(string content)
         {
-
+            return null;
         }
         public static string ExportKeyValue(KeyValue kv)
         {
             StringBuilder sb = new StringBuilder();
             NodeToFullFormatString(kv.Root, sb);
             Console.Write(sb.ToString());
+            return sb.ToString();
         }
         private static void NodeToFullFormatString(KVNode node, StringBuilder sb, int depth=0)
         {
