+++ Valve/Valve.KeyValue/Properties/AssemblyInfo.cs
