+++ Valve/Valve.KeyValue/Valve.KeyValue.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{F1FD06B9-DF2F-4336-83F4-6A1C6372AADF}</ProjectGuid>
-    <OutputType>Exe</OutputType>
+    <ProjectGuid>{66A8409E-243D-462B-A6D0-DE782B6FABA7}</ProjectGuid>
+    <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Valve.VTF</RootNamespace>
-    <AssemblyName>valve.fileio</AssemblyName>
+    <RootNamespace>Valve.KeyValue</RootNamespace>
+    <AssemblyName>valve</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -29,9 +29,6 @@
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
-  <PropertyGroup>
-    <StartupObject />
-  </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
@@ -43,9 +40,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="KeyValue.cs" />
+    <Compile Include="KVNode.cs" />
+    <Compile Include="KVFile.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="VMT.cs" />
-    <Compile Include="VTF.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
