+++ Valve/Valve.VTF/Properties/AssemblyInfo.cs
