+++ Valve/Valve.VTF/VTF.cs
