+++ Valve/Valve.VTF/Valve.VTF.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{BB0EEF20-A25E-499A-B360-0C80E0F651A4}</ProjectGuid>
-    <OutputType>Library</OutputType>
+    <ProjectGuid>{F1FD06B9-DF2F-4336-83F4-6A1C6372AADF}</ProjectGuid>
+    <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>TF2.Info</RootNamespace>
-    <AssemblyName>tf2.info</AssemblyName>
+    <RootNamespace>Valve.VTF</RootNamespace>
+    <AssemblyName>valve</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -29,6 +29,9 @@
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
+  <PropertyGroup>
+    <StartupObject />
+  </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
@@ -41,8 +44,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TFClass.cs" />
-    <Compile Include="TFSlot.cs" />
+    <Compile Include="VMT.cs" />
+    <Compile Include="VTF.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
